+++ site/src/routes/_layout.svelte
@@ -24,7 +24,7 @@
 
 		<NavItem external="https://sapper.svelte.dev">Sapper</NavItem>
 
-		<NavItem external="https://discord.gg/yy75DKs" title="Discord Chat">
+		<NavItem external="chat" title="Discord Chat">
 			<Icon name="message-square"/>
 		</NavItem>