+++ test/js/samples/bind-open/input.svelte
@@ -1,0 +1,7 @@
+<script>
+	export let open;
+</script>
+
+<details bind:open={open}>
+	<summary>summary</summary>content
+</details>