+++ test/js/samples/media-bindings/expected.js
@@ -8,6 +8,7 @@ import {
 	insert,
 	listen,
 	noop,
+	raf,
 	run_all,
 	safe_not_equal,
 	time_ranges_to_array
@@ -18,7 +19,7 @@ function create_fragment(ctx) {
 
 	function audio_timeupdate_handler() {
 		cancelAnimationFrame(audio_animationframe);
-		if (!audio.paused) audio_animationframe = requestAnimationFrame(audio_timeupdate_handler);
+		if (!audio.paused) audio_animationframe = raf(audio_timeupdate_handler);
 		audio_updating = true;
 		ctx.audio_timeupdate_handler.call(audio);
 	}