+++ test/runtime/samples/component-namespaced/components.js
@@ -1,0 +1,3 @@
+import Foo from './Foo.svelte';
+
+export default { Foo };