+++ test/runtime/samples/element-invalid-name/_config.js
@@ -2,4 +2,4 @@ export default {
 	html: `
 		<foo-bar>Hello</foo-bar>
 	`
-}
+};