+++ test/runtime/samples/store-imported-module-b/foo.js
@@ -1,3 +1,3 @@
-import { writable } from '../../../../store.js';
+import { writable } from '../../../../store';
 
 export default writable(42);